Add unit tests for NativeLibraryInterop.cs
While implementing these tests, I decided to change a few components of the native types. Notably, I replaced most linked-list structures with pointer/count arrays because linked-lists seemed like they would be inconvenient to work with. Additionally, I had to make a NativeBool enum in C# because the built-in bool type is not blittable.