Write the graph optimizer
After program graph construction, the optimizer runs. It performs two main functions:

(1) Removing nodes which don't contribute to the output and deduplicating logic. For example, "(x + y) * (x + y)" will effectively get transformed into "z = x + y; z * z" (this runs at the graph level, not on source code).

(2) Applying optimization rule. An optimization rule consists of an input pattern, which is detected in the graph (e.g. "x + -y") and a simplified output pattern which the input pattern is replaced with (e.g. "x - y").

The second function, optimization rule application, is not quite trivial and the high level algorithm is described here. The behavior we seek is that optimization rules are to be repeatedly applied at the earliest possible point in the graph until no more apply. However, re-traversing the entire graph every time an optimization rule is applied has O(n^2) complexity. Instead, we take a more optimized approach with identical end results. A BFS traversal of the graph is performed. When each node is visited, an optimized tree structure is used to determine if any optimization rules can be applied using that node as the rule's root. If multiple optimization rules apply, the "most complex" rule is chosen (currently, a simple complexity measure is used, which could be improved in the future). The optimization rule is then applied by swapping out the nodes in the rule's input pattern for newly generated nodes in the rule's output pattern. After an optimization rule is applied, it is possible that a rule could be matched further back in the graph. To avoid missing this case, we rewind the BFS traversal to a point which could not possibly miss any optimization rule application opportunities. This is done by first calculating the maximum depth of each native module call within any optimization rule. When new native module call nodes are added to the graph, we rewind the BFS traversal from at least (new node depth) - (max native module call depth within any rule). (There are some additional edge-case details in this step explained in the code).