Add an optimization rule cycle detector
It is possible to write optimization rules which will get the optimizer stuck in a loop (e.g. A -> B -> C -> A). This is now detected by taking a hash of the graph at increasing intervals. Graph hashes can be compared to determine if two graphs are identical. Eventually, if there is a cycle, the interval of that cycle (or a multiple of it) will be tested and the cycle will be detected by noticing that the graph's hash is identical to a previous hash value.