Native modules now output one latency value per output rather than a single latency value
This change was made so that latency values can be specified in terms of the parameter upsample factor rather than the native module upsample factor. This will be important when implementing resampler modules whose filter latency will be in terms of the target sample rate. Also fixed a minor bug having to do with output argument source locations when issuing native module calls.