Add engine code to deserialize program from byte array
Additionally, an SHA256 of the program file's content is now included in the header for validation. This is done so that the engine can simply check the hash to verify that the contents of the program were not changed after compilation, rather than performing explicit validation of the program graph. Full graph validation is probably more robust but it's a lot more code and duplicated logic (across the compiler and engine) which isn't currently necessary.