Add unit tests for ConstantManager
To unit test this class, program graph nodes must be constructed manually. Previously this was impossible because program graph nodes were friended with the Program class which would modify node internals after nodes were created (this was done so that the nodes themselves didn't need to have setters exposed for internal details). To get around this, a new class, ProgramGraphNodeModifier, was created. Nodes are now friends with this class instead of Program, and Program invokes the functions in this class to do the same modifications that it was doing directly before. ProgramGraphNodeModifier is also now used by unit tests so that node graphs can be manually built.